+++ src/player.h
@@ -3,15 +3,15 @@
 
 typedef struct {
     char name[100];
-    size_t health;
-    size_t health_max;
-    size_t strength;
+    unsigned health;
+    unsigned health_max;
+    unsigned strength;
     inventory_t inventory;
 } player_t;
 
-void player_init(player_t *p, const char *name, size_t hp, size_t str);
-void player_take_damage(player_t *p, size_t damage);
-void player_heal(player_t *p, size_t amount);
+void player_init(player_t *p, const char *name, unsigned hp, unsigned str);
+void player_take_damage(player_t *p, unsigned damage);
+void player_heal(player_t *p, unsigned amount);
 
-size_t player_max_capacity(const player_t *p);
+unsigned player_max_capacity(const player_t *p);
 bool player_is_inventory_overloaded(const player_t *p);
